airflow/requirements.txt
apache-airflow-providers-google
pyarrow
pandas
dbt-bigquery